requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
crispy-bulma==0.11.0
dj-database-url==2.1.0
Django==5.0.4
django-crispy-forms==2.1
django-extensions==3.2.3
django-formtools==2.3
django-widget-tweaks==1.5.0
psycopg2-binary==2.9.9
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
